requirements.txt
asgiref==3.8.1
boto3==1.35.11
botocore==1.35.11
crispy-bootstrap5==2024.2
django-allauth==64.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-storages==1.14.4
gunicorn==23.0.0
jmespath==1.0.1
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2